requirements.txt
# Bitrecs Subnet Requirements
# Core Bittensor dependencies
bittensor==9.9.0
bittensor-cli==9.10.1
substrate-interface==1.7.11
cryptography==45.0.4

# Web framework and API
fastapi==0.104.1
starlette==0.37.2
uvicorn==0.24.0
pydantic==2.10.4

# Data processing and ML
numpy==2.0.2
pandas==2.2.3

# LLM and AI
openai==1.93.0
tiktoken==0.9.0

# JSON and data validation
json-repair==0.47.6
jsonschema==4.24.0

# Environment and configuration
python-dotenv==1.0.1

# HTTP requests
requests==2.31.0

# Async and concurrency
anyio==4.0.0

# Monitoring and logging
wandb==0.20.1

# Rate limiting
slowapi==0.1.9

# Development and testing (optional)
pytest==8.3.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code formatting (optional)
black==23.11.0
isort==5.12.0

# Additional utilities
setuptools>=70.0.0
wheel>=0.41.0
